@@ -3,7 +3,7 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
 main() {
-  PreferencesService _preferencesService;
+  late PreferencesService _preferencesService;
 
   setUp(() async {
     SharedPreferences.setMockInitialValues({'testKey': 'testValue'});
