@@ -1,5 +1,6 @@
 import 'package:brick_app/model/brick_set.dart';
 import 'package:brick_app/model/rebrickable_model.dart';
+import 'package:brick_app/model/set_or_moc.dart';
 import 'package:brick_app/widgets/brick_app_bar.dart';
 import 'package:brick_app/widgets/sets_grid_view.dart';
 import 'package:flutter/material.dart';
@@ -16,12 +17,12 @@ class MocPage extends StatelessWidget {
       appBar: BrickAppBar(
         Text(brickSet.name),
       ),
-      body: FutureBuilder(
+      body: FutureBuilder<List<SetOrMoc>>(
         future: service.getMocsFromSet(setNum: brickSet.setNum),
         builder: (context, snapshot) {
           if (snapshot.hasData) {
             return SetsGridView(
-              snapshot.data,
+              snapshot.data!,
               withButtons: false,
             );
           } else {
